+++ example/bin/main.dart
@@ -1,7 +1,9 @@
-import 'package:example/injector.dart';
+import 'package:example/injector/injector.dart';
 import 'package:example/services/register_module.dart';
+import 'package:injectable/injectable.dart';
 
 void main(List<String> arguments) {
-  configureDependencies();
+  configureDependencies(Environment.dev);
   print(getIt<Client>().url);
 }
+