+++ example/lib/injector/injector.dart
@@ -1,0 +1,12 @@
+import 'package:get_it/get_it.dart';
+import 'package:injectable/injectable.dart';
+
+import 'injector.config.dart';
+
+/// getIt instance that's gonna
+/// be used throughout the App
+final GetIt getIt = GetIt.instance;
+
+/// entry point for injection
+@InjectableInit()
+void configureDependencies(String environment) => $initGetIt(getIt, environment: environment);