+++ injectable_generator/lib/injectable_config_generator.dart
@@ -33,7 +33,8 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
     jsonData.forEach((json) => deps.add(DependencyConfig.fromJson(json)));
 
     _reportMissingDependencies(deps);
-    return ConfigCodeGenerator(deps).generate();
+    _reportDuplicateDependencies(deps);
+    return ConfigCodeGenerator(deps, element.source.uri.path).generate();
   }
 
   void _reportMissingDependencies(List<DependencyConfig> deps) {
@@ -51,9 +52,23 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
     });
 
     if (messages.isNotEmpty) {
-      messages.add(
-          '\nDid you forget to annotate the above class(es) or their implementation with @injectable?');
+      messages.add('\nDid you forget to annotate the above classe(s) or their implementation with @injectable?');
       printBoxed(messages.join('\n'));
+    }
+  }
+
+  void _reportDuplicateDependencies(List<DependencyConfig> deps) {
+    final registeredDeps = <DependencyConfig>[];
+    for (var dep in deps) {
+      var registered = registeredDeps.where((elm) => elm.type == dep.type && elm.instanceName == dep.instanceName);
+      if (registered.isEmpty) {
+        registeredDeps.add(dep);
+      } else {
+        Set<String> registeredEnvironments = registered.fold(<String>{}, (prev, elm) => prev..addAll(elm.environments));
+        if (registeredEnvironments.isEmpty || dep.environments.any((env) => registeredEnvironments.contains(env))) {
+          throwBoxed('${dep.typeImpl} [${dep.type}] env: ${dep.environments} \nis registered more than once under the same environment');
+        }
+      }
     }
   }
 }